+++ models/diaryEntryFood.go
@@ -1,9 +1,9 @@
 package models
 
 type DiaryEntryFood struct {
-	DiaryEntryFoodID int
-	DiaryEntryID     int
-	FoodID           int
+	DiaryEntryFoodID int64
+	DiaryEntryID     int64
+	FoodID           int64
 	Servings         float32
 	Calories         float32
 	Fat              float32
