+++ models/diaryEntryRecipe.go
@@ -1,7 +1,7 @@
 package models
 
 type DiaryEntryRecipe struct {
-	DiaryEntryRecipeID int
-	DiaryEntryID       int
-	RecipeID           int
+	DiaryEntryRecipeID int64
+	DiaryEntryID       int64
+	RecipeID           int64
 }
